+++ apps/examples/src/misc/develop.tsx
@@ -2,17 +2,23 @@ import { getLicenseKey } from '@tldraw/dotcom-shared'
 import {
 	DefaultContextMenu,
 	DefaultContextMenuContent,
+	DefaultDebugMenu,
+	DefaultDebugMenuContent,
+	ExampleDialog,
 	TLComponents,
 	Tldraw,
 	TldrawUiMenuActionCheckboxItem,
 	TldrawUiMenuActionItem,
 	TldrawUiMenuGroup,
+	TldrawUiMenuItem,
 	track,
+	useDialogs,
 	useEditor,
 } from 'tldraw'
 import 'tldraw/tldraw.css'
 import { trackedShapes, useDebugging } from '../hooks/useDebugging'
 import { usePerformance } from '../hooks/usePerformance'
+import { A11yResultTable } from './a11y'
 import { getDiff } from './diff'
 
 const ContextMenu = track(() => {
@@ -38,8 +44,47 @@ const ContextMenu = track(() => {
 	)
 })
 
+function A11yAudit() {
+	const { addDialog } = useDialogs()
+
+	const runA11yAudit = async () => {
+		const axe = (await import('axe-core')).default
+		axe.run(document, {}, (err, results) => {
+			if (err) throw err
+
+			// eslint-disable-next-line no-console
+			console.debug('[a11y]:', results)
+
+			addDialog({
+				component: ({ onClose }) => (
+					<ExampleDialog
+						body={<A11yResultTable results={results} />}
+						title="Accessibility Audit Results"
+						maxWidth="80vw"
+						cancel="Close"
+						confirm="Ok"
+						onCancel={() => onClose()}
+						onContinue={() => onClose()}
+					/>
+				),
+				onClose: () => {
+					void null
+				},
+			})
+		})
+	}
+
+	return <TldrawUiMenuItem id="a11y-audit" onSelect={runA11yAudit} label={'A11y audit'} />
+}
+
 const components: TLComponents = {
 	ContextMenu,
+	DebugMenu: () => (
+		<DefaultDebugMenu>
+			<A11yAudit />
+			<DefaultDebugMenuContent />
+		</DefaultDebugMenu>
+	),
 }
 
 function afterChangeHandler(prev: any, next: any) {
