+++ packages/tldraw/package.json
@@ -78,7 +78,6 @@
 		"@types/classnames": "^2.3.4",
 		"@types/lz-string": "^1.5.0",
 		"@types/react": "^18.3.18",
-		"axe-core": "^4.10.3",
 		"chokidar-cli": "^3.0.0",
 		"jest-canvas-mock": "^2.5.2",
 		"jest-environment-jsdom": "^29.7.0",
